v1/requirements.txt
flash
flask
flask-mysqldb
flask-sqlalchemy
mysqlclient
python-dotenv
sqlalchemy
flask-login
